- Implemented ACHIEVEMENT_CRITERIA_TYPE_LOOT_ITEM, retroactive as usual. - Fixed compile-time warnings in Unit.cpp.
git-svn-id: svn://svn.aspiredev.org/AspireCore/branches/Hearthstone@545 a6a5f009-272a-4b40-a74d-5f9816a51f88

+++ src/hearthstone-world/Unit.cpp
@@ -6410,7 +6410,7 @@ void Unit::SetPowerType(uint8 type)
 	SetByte(UNIT_FIELD_BYTES_0,3,type);
 	if (type == POWER_TYPE_RUNIC)
 	{
-		SetUInt32Value(PLAYER_RUNE_REGEN_1, 0.100000);
-		SetUInt32Value(PLAYER_RUNE_REGEN_1_1, 0.100000);
+		SetFloatValue(PLAYER_RUNE_REGEN_1, 0.100000f);
+		SetFloatValue(PLAYER_RUNE_REGEN_1_1, 0.100000f);
 	}
 }
